# contracts/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nektron Moments API",
    "version": "0.1.0",
    "description": "Versioned foundation contract for Nektron Moments mobile, CLI, and future native Windows clients. The v1 API stores metadata for Local sources, manages exact-deduplicated originals for Remote sources, and exposes cursor-based media evidence without coupling clients to the MySQL schema."
  },
...
  "security": [
    {
      "CognitoBearer": []
    }
  ],
  "paths": {
...
    "/v1/me": {
      "get": {
        "tags": [
          "Account"
        ],
        "operationId": "getCurrentUser",
        "summary": "Get the signed-in Nektron Moments user",
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/devices": {
      "get": {
        "tags": [
          "Account"
        ],
        "operationId": "listDevices",
        "summary": "List registered devices",
        "parameters": [
          {
            "$ref": "#/components/parameters/CursorQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Account"
        ],
        "operationId": "registerDevice",
        "summary": "Register or refresh an app installation",
        "description": "The stable client installation ID makes registration safely repeatable across app restarts.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistrationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing device registration refreshed.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/IdempotencyReplayedHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "201": {
            "description": "Device registered.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/IdempotencyReplayedHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1/sources": {
      "get": {
        "tags": [
          "Sources"
        ],
        "operationId": "listSources",
        "summary": "List the user's media sources",
        "parameters": [
          {
            "$ref": "#/components/parameters/CursorQuery"
          },
          {
            "$ref": "#/components/parameters/LimitQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaSourcePage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "tags": [
          "Sources"
        ],
        "operationId": "createSource",
        "summary": "Create a Local or Remote media source",
        "description": "Storage mode defaults to Local. A source belongs to exactly one registered device. Recreating a previously removed source reactivates the same source identity.",
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaSourceCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing compatible source returned or removed source reactivated.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/IdempotencyReplayedHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaSource"
                }
              }
            }
          },
          "201": {
            "description": "Source created.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              },
              "Idempotency-Replayed": {
                "$ref": "#/components/headers/IdempotencyReplayedHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaSource"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Unprocessable"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/v1/uploads/{uploadSessionId}": {
      "get": {
        "tags": [
          "Uploads"
        ],
        "operationId": "getUploadSession",
        "summary": "Get resumable upload state",
        "description": "Returns the authoritative uploaded-part list and current lease state after an app restart or interrupted transfer.",
        "parameters": [
          {
            "$ref": "#/components/parameters/UploadSessionIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Current upload session state.",
            "headers": {
              "X-Request-Id": {
                "$ref": "#/components/headers/RequestIdHeader"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadSessionStatus"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CognitoBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Amazon Cognito access token for the Nektron Moments user pool."
      },
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "The same Cognito access token with the Nektron Moments administrator claim."
      }
    },
    "headers": {
      "RequestIdHeader": {
        "description": "Server-generated request correlation identifier.",
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128
        }
      },
...
    },
    "parameters": {
...
      "LimitQuery": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 200,
          "default": 50
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication or provider signature verification failed.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "An unexpected server error occurred.",
        "headers": {
          "X-Request-Id": {
            "$ref": "#/components/headers/RequestIdHeader"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "CurrentUser": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "email",
          "createdAtUtc"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "displayName": {
            "type": "string",
            "maxLength": 200,
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}